Refactor chat interfaces into chat.go from interface.go
Consolidated the `HistoryService` and `Service` interfaces into `chat.go` for improved file organization and maintainability. Removed the now redundant `interface.go` file. This change simplifies the code structure and centralizes chat-related definitions.
